Time: 0 ms (100%), Space: 134.4 MB (64.44%) - LeetHub

+++ 2090-k-radius-subarray-averages/2090-k-radius-subarray-averages.cpp
@@ -6,7 +6,7 @@ public:
 
         int size = nums.size();
         
-        vector<long long> ans (size, -1);
+        vector<int> ans (size, -1);
         
         if(size < 2*k + 1)
             return ans;
